chore: split collection layout into client and server
